Move options to Mesh rather than MeshFilter
