+++ openmc/filter.py
@@ -4,7 +4,6 @@ from collections.abc import Iterable
 import hashlib
 from itertools import product
 from numbers import Real, Integral
-from typing import Optional
 import warnings
 
 import lxml.etree as ET
@@ -796,9 +795,6 @@ class MeshFilter(Filter):
     ----------
     mesh : openmc.MeshBase
         The mesh object that events will be tallied onto
-    options : str, optional
-        Special options that control spatial search data structures used. This
-        is currently only used to set parameters for MOAB's AdaptiveKDTree.
     filter_id : int
         Unique identifier for the filter
 
@@ -806,9 +802,6 @@ class MeshFilter(Filter):
     ----------
     mesh : openmc.MeshBase
         The mesh object that events will be tallied onto
-    options : str
-        Special options that control spatial search data structures used. This
-        is currently only used to set parameters for MOAB's AdaptiveKDTree.
     id : int
         Unique identifier for the filter
     translation : Iterable of float
@@ -822,9 +815,8 @@ class MeshFilter(Filter):
 
     """
 
-    def __init__(self, mesh, options=None, filter_id=None):
+    def __init__(self, mesh, filter_id=None):
         self.mesh = mesh
-        self.options = options
         self.id = filter_id
         self._translation = None
 
@@ -854,12 +846,8 @@ class MeshFilter(Filter):
         mesh_id = group['bins'][()]
         mesh_obj = kwargs['meshes'][mesh_id]
         filter_id = int(group.name.split('/')[-1].lstrip('filter '))
-        if 'options' in group.attrs:
-            options = group.attrs['options'].decode()
-        else:
-            options = None
 
-        out = cls(mesh_obj, options=options, filter_id=filter_id)
+        out = cls(mesh_obj, filter_id=filter_id)
 
         translation = group.get('translation')
         if translation:
@@ -882,15 +870,6 @@ class MeshFilter(Filter):
                 self.bins = []
         else:
             self.bins = list(mesh.indices)
-
-    @property
-    def options(self) -> Optional[str]:
-        return self._options
-
-    @options.setter
-    def options(self, options: Optional[str]):
-        cv.check_type('filter options', options, (str, type(None)))
-        self._options = options
 
     @property
     def shape(self):
@@ -988,8 +967,6 @@ class MeshFilter(Filter):
         """
         element = super().to_xml_element()
         element[0].text = str(self.mesh.id)
-        if self.options is not None:
-            element.set('options', self.options)
         if self.translation is not None:
             element.set('translation', ' '.join(map(str, self.translation)))
         return element
@@ -999,8 +976,7 @@ class MeshFilter(Filter):
         mesh_id = int(get_text(elem, 'bins'))
         mesh_obj = kwargs['meshes'][mesh_id]
         filter_id = int(elem.get('id'))
-        options = elem.get('options')
-        out = cls(mesh_obj, options=options, filter_id=filter_id)
+        out = cls(mesh_obj, filter_id=filter_id)
 
         translation = elem.get('translation')
         if translation:
